+++ NEWS.md
@@ -13,6 +13,8 @@
 - New functions `pl$datetime()`, `pl$date()`, and `pl$time()` to easily create
   Expr of class datetime, date, and time via columns and literals (#918).
 - New function `pl$arg_where()` to get the indices that match a condition (#922).
+- New method `<LazyFrame>$show_graph()` to plot the query plan of a LazyFrame 
+  (#928).
 
 ## Polars R Package 0.15.1